requirements.txt
asn1crypto==0.24.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.3
django-after-response==0.2.2
django-bootstrap4==0.0.7
django-filter==2.1.0
djangorestframework==3.9.2
emoji-data-python==1.1.0
idna==2.8
Markdown==3.0.1
markdown2==2.3.7
munch==2.3.2
psycopg2-binary==2.8.2
pycparser==2.19
python-slugify==1.2.6
pytz==2018.9
requests==2.21.0
six==1.12.0
slackclient==1.3.1
sqlparse==0.3.0
statuspageio==0.0.1
Unidecode==1.0.23
urllib3==1.24.3
webencodings==0.5.1
websocket-client==0.54.0
boto3==1.9.57
mysqlclient==1.4.2.post1
opsgenie-sdk==2.0.1
